Prod-0: Move data model objects to dal folder into their appropriate pacakge
