+++ src/main/java/com/indiraactive/fulfillmentplatform/dal/scriptRunAuditEntry/ScriptRunAuditEntryRepository.java
@@ -1,8 +1,6 @@
 package com.indiraactive.fulfillmentplatform.dal.scriptRunAuditEntry;
 
-import com.indiraactive.fulfillmentplatform.model.db.ScriptRunAuditEntry;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
