+++ src/main/java/com/indiraactive/fulfillmentplatform/factory/ScheduledTaskJpaFactory.java
@@ -2,10 +2,10 @@ package com.indiraactive.fulfillmentplatform.factory;
 
 import com.indiraactive.fulfillmentplatform.dal.supplier.SupplierRepository;
 import com.indiraactive.fulfillmentplatform.model.ScheduledTask;
-import com.indiraactive.fulfillmentplatform.model.ScheduledTaskJpa;
-import com.indiraactive.fulfillmentplatform.model.ScheduledTaskRunDaysJpa;
-import com.indiraactive.fulfillmentplatform.model.ScheduledTaskRunTimeJpa;
-import com.indiraactive.fulfillmentplatform.model.db.Supplier;
+import com.indiraactive.fulfillmentplatform.dal.scheduledTask.ScheduledTaskJpa;
+import com.indiraactive.fulfillmentplatform.dal.scheduledTask.ScheduledTaskRunDaysJpa;
+import com.indiraactive.fulfillmentplatform.dal.scheduledTask.ScheduledTaskRunTimeJpa;
+import com.indiraactive.fulfillmentplatform.dal.supplier.Supplier;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
