+++ src/main/java/com/indiraactive/fulfillmentplatform/service/SchedulerImpl.java
@@ -3,8 +3,8 @@ package com.indiraactive.fulfillmentplatform.service;
 import com.indiraactive.fulfillmentplatform.dal.scheduledTask.ScheduledTaskRepository;
 import com.indiraactive.fulfillmentplatform.dal.scriptRunAuditEntry.ScriptRunAuditEntryRepository;
 import com.indiraactive.fulfillmentplatform.dal.supplier.SupplierRepository;
-import com.indiraactive.fulfillmentplatform.model.ScheduledTaskJpa;
-import com.indiraactive.fulfillmentplatform.model.ScheduledTaskRunDaysJpa;
+import com.indiraactive.fulfillmentplatform.dal.scheduledTask.ScheduledTaskJpa;
+import com.indiraactive.fulfillmentplatform.dal.scheduledTask.ScheduledTaskRunDaysJpa;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
