+++ src/main/java/com/indiraactive/fulfillmentplatform/viewModel/RunHistoryViewModel.java
@@ -1,8 +1,8 @@
 package com.indiraactive.fulfillmentplatform.viewModel;
 
 import com.indiraactive.fulfillmentplatform.model.RunHistoryModel;
-import com.indiraactive.fulfillmentplatform.model.db.ScriptRunAuditEntry;
-import com.indiraactive.fulfillmentplatform.model.db.Supplier;
+import com.indiraactive.fulfillmentplatform.dal.scriptRunAuditEntry.ScriptRunAuditEntry;
+import com.indiraactive.fulfillmentplatform.dal.supplier.Supplier;
 
 import java.util.Date;
 import java.util.LinkedList;
